app/requirements.txt
Django>=3.2,<4.0
gunicorn
faiss-cpu
sentence-transformers
pymongo
bs4
pyspark
lxml
html5lib
matplotlib
django-cors-headers
torch
torchvision
pillow


